@@ -22,57 +22,61 @@
   "author": "vmgabriel, dbgroldan",
   "license": "ISC",
   "dependencies": {
+    "@hapi/boom": "^9.1.0",
+    "@hapi/joi": "^17.1.1",
     "bcrypt": "^4.0.1",
     "body-parser": "^1.19.0",
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
-    "couchbase": "^2.6.11",
+    "couchbase": "^2.6.12",
     "dotenv": "^8.2.0",
     "envalid": "^6.0.1",
     "express": "^4.17.1",
     "helmet": "^3.22.0",
     "https": "^1.0.0",
+    "joigoose": "^7.0.0",
     "jwt-simple": "^0.5.6",
     "method-override": "^3.0.0",
-    "moment": "^2.25.1",
+    "moment": "^2.26.0",
+    "mongoose": "^5.9.15",
     "morgan": "^1.10.0",
     "passport": "^0.4.1",
     "passport-cookie": "^1.0.6",
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
     "ramda": "^0.27.0",
-    "tsoa": "^3.0.7",
-    "uuid": "^8.0.0",
-    "@hapi/boom": "^9.1.0",
-    "@hapi/joi": "^15.0.3"
+    "tsoa": "^3.0.8",
+    "uuid": "^8.1.0"
   },
   "devDependencies": {
-    "grunt-ts": "^6.0.0-beta.22",
-    "grunt-contrib-watch": "^1.1.0",
-    "grunt-typedoc": "^0.2.4",
-    "@types/cookie-parser": "^1.4.2",
-    "@types/method-override": "0.0.31",
-    "@types/passport": "^1.0.3",
     "@types/bcrypt": "^3.0.0",
+    "@types/cookie-parser": "^1.4.2",
     "@types/cors": "^2.8.6",
     "@types/express": "^4.17.6",
+    "@types/hapi__joi": "^17.1.0",
     "@types/helmet": "0.0.46",
+    "@types/method-override": "0.0.31",
     "@types/mocha": "^7.0.2",
-    "@types/node": "^13.13.4",
+    "@types/mongoose": "^5.7.21",
+    "@types/node": "^13.13.9",
+    "@types/passport": "^1.0.3",
     "@types/passport-jwt": "^3.0.3",
     "@types/passport-local": "^1.0.33",
-    "@types/ramda": "^0.27.4",
+    "@types/ramda": "^0.27.6",
     "@types/swagger-ui-express": "^4.1.2",
-    "@types/uuid": "^7.0.3",
+    "@types/uuid": "^7.0.4",
     "grunt": "^1.1.0",
     "grunt-browserify": "^5.3.0",
     "grunt-cli": "^1.3.2",
-    "mocha": "^7.1.2",
+    "grunt-contrib-watch": "^1.1.0",
+    "grunt-ts": "^6.0.0-beta.22",
+    "grunt-typedoc": "^0.2.4",
+    "mocha": "^7.2.0",
     "proxyquire": "^2.1.3",
     "sinon": "^9.0.2",
     "supertest": "^4.0.2",
-    "ts-node": "^8.9.1",
-    "typedoc": "^0.17.6",
-    "typescript": "^3.8.3"
+    "ts-node": "^8.10.1",
+    "typedoc": "^0.17.7",
+    "typescript": "^3.9.3"
   }
 }
