@@ -1,11 +1,14 @@
 // Develop vmgabriel
 
+// interfaces
+import { IAbstract } from './abtract';
+
+export const nameTable = 'module';
+export const stateName = 'moduleIsValid';
+
 /** Module Interface  */
-export interface IModule {
-  moduleId: string;
+export interface IModule extends IAbstract {
   moduleName: string;
   moduleDescription: string;
-  moduleLink: string;
-  modulePermission: Array<string>;
   moduleIsValid: boolean;
 }
