@@ -11,6 +11,8 @@ export class ModuleModel extends AbstractModel {
   private moduleId: any;
   private moduleNameSchema: any;
   private moduleDescriptionSchema: any;
+  private moduleMetaInfo: any
+  private moduleIsValid: any;
 
   constructor() {
     super();
@@ -18,13 +20,16 @@ export class ModuleModel extends AbstractModel {
     this.moduleId = joi.string();
     this.moduleNameSchema = joi.string().max(80);
     this.moduleDescriptionSchema = joi.string().max(200);
+    this.moduleIsValid = joi.boolean();
+    this.moduleMetaInfo = joi.any();
   }
 
   /** get Scheme Created  */
   public getCreateScheme(): any {
     return {
       moduleName: this.moduleNameSchema.required(),
-      moduleDescription: this.moduleDescriptionSchema.required()
+      moduleDescription: this.moduleDescriptionSchema.required(),
+      moduleMetaInfo: this.moduleMetaInfo
     };
   }
 
@@ -32,7 +37,18 @@ export class ModuleModel extends AbstractModel {
   public getUpdateScheme() {
     return {
       moduleName: this.moduleNameSchema,
-      moduleDescription: this.moduleDescriptionSchema
+      moduleDescription: this.moduleDescriptionSchema,
+      moduleMetaInfo: this.moduleMetaInfo
+    };
+  }
+
+  public getData() {
+    return {
+      moduleName: this.moduleNameSchema.required(),
+      moduleDescription: this.moduleDescriptionSchema.required(),
+      moduleMetaInfo: this.moduleMetaInfo,
+      moduleIsValid: this.moduleIsValid.required(),
+      deletedAt: this.deletedAt
     };
   }
 
