@@ -16,11 +16,17 @@ import {
   IAndOrFilter,
   IFromFilter,
 } from '../interfaces/filter';
+import { RoleMongoRepository } from '../libs/mongo/Repositories/mongoRoleRepository';
 
 /** Role Service  */
 export class RoleService extends AbstractService {
   constructor() {
-    super('role', new RoleModel(), 'roleIsValid');
+    super(
+      'role',
+      new RoleModel(),
+      'roleIsValid',
+      new RoleMongoRepository()
+    );
   }
 
   /**
@@ -35,79 +41,79 @@ export class RoleService extends AbstractService {
     filter: IAndOrFilter,
     limit: number = 10,
     offset: number = 0
-  ) {
-    if (attributes.length > 0) {
-      attributes = R.map((d: IAttributeChange) => { return {
-        name: 'rol',
-        as: d.as,
-        column: d.column
-      }; }, attributes);
-    } else {
-      attributes.push({
-        as: '',
-        name: 'rol',
-        column: '*'
-      });
-    }
+  ) {}
+  //   if (attributes.length > 0) {
+  //     attributes = R.map((d: IAttributeChange) => { return {
+  //       name: 'rol',
+  //       as: d.as,
+  //       column: d.column
+  //     }; }, attributes);
+  //   } else {
+  //     attributes.push({
+  //       as: '',
+  //       name: 'rol',
+  //       column: '*'
+  //     });
+  //   }
 
-    attributes.push({
-      as: '',
-      name: '',
-      column: 'permissions'
-    });
+  //   attributes.push({
+  //     as: '',
+  //     name: '',
+  //     column: 'permissions'
+  //   });
 
-    const joins: IFromFilter = {
-      from: [
-        {
-          name: 'rol',
-          default: true
-        },
-        {
-          name: 'rolePermission',
-          joinType: 'LEFT',
-          unionType: 'NEST',
-          onType: 'KEYS',
-          onValue: 'rol.rolePermission[*].roleModuleId',
-          default: false
-        }
-      ]
-    };
+  //   const joins: IFromFilter = {
+  //     from: [
+  //       {
+  //         name: 'rol',
+  //         default: true
+  //       },
+  //       {
+  //         name: 'rolePermission',
+  //         joinType: 'LEFT',
+  //         unionType: 'NEST',
+  //         onType: 'KEYS',
+  //         onValue: 'rol.rolePermission[*].roleModuleId',
+  //         default: false
+  //       }
+  //     ]
+  //   };
 
-    let otherContent = `LET permissions = ARRAY {
-        "roleModulePermission": perm.roleModulePermission,
-        "roleModuleId": perm.roleModuleId,
-        "module": IFNULL(
-          FIRST rm FOR rm IN rolePermission
-          WHEN META(rm).id = perm.roleModuleId END, MISSING
-        )
-      } FOR perm IN rol.rolePermission END`;
+  //   let otherContent = `LET permissions = ARRAY {
+  //       "roleModulePermission": perm.roleModulePermission,
+  //       "roleModuleId": perm.roleModuleId,
+  //       "module": IFNULL(
+  //         FIRST rm FOR rm IN rolePermission
+  //         WHEN META(rm).id = perm.roleModuleId END, MISSING
+  //       )
+  //     } FOR perm IN rol.rolePermission END`;
 
-    return new Promise(async (resolve: any, reject: any) => {
-      try {
-        const { query, count }: IResponseFilterDb = await this.connection.advancedFilter(
-          this.collection,
-          this.attributeState,
-          attributes,
-          joins,
-          otherContent,
-          filter,
-          limit,
-          offset
-        );
-        const length = await count;
+  //   return new Promise(async (resolve: any, reject: any) => {
+  //     try {
+  //       const { query, count }: IResponseFilterDb = await this.connection.advancedFilter(
+  //         this.collection,
+  //         this.attributeState,
+  //         attributes,
+  //         joins,
+  //         otherContent,
+  //         filter,
+  //         limit,
+  //         offset
+  //       );
+  //       const length = await count;
 
-        const message = {
-          code: 200,
-          count: length,
-          rows: query
-        };
+  //       const message = {
+  //         code: 200,
+  //         count: length,
+  //         rows: query
+  //       };
 
-        resolve(message);
-      } catch (err) {
-        reject(err);
-      }
-    });
-  }
+  //       resolve(message);
+  //     } catch (err) {
+  //       reject(err);
+  //     }
+  //   });
+  // }
 
   // End Class
 }
